+++ admin/src/scenes/missions/list.jsx
@@ -64,24 +64,6 @@ export default function List() {
 
   async function transform(data, selectedFields) {
     let all = data;
-    if (selectedFields.includes("tutor")) {
-      const tutorIds = [...new Set(data.map((item) => item.tutorId).filter((e) => e))];
-      if (tutorIds?.length) {
-        const { responses } = await api.esQuery("referent", { size: ES_NO_LIMIT, query: { ids: { type: "_doc", values: tutorIds } } });
-        if (responses.length) {
-          const tutors = responses[0]?.hits?.hits.map((e) => ({ _id: e._id, ...e._source }));
-          all = data.map((item) => ({ ...item, tutor: tutors?.find((e) => e._id === item.tutorId) }));
-        }
-      }
-    }
-    if (["structureInfo", "structureLocation"].some((e) => selectedFields.includes(e))) {
-      const structureIds = [...new Set(data.map((item) => item.structureId).filter((e) => e))];
-      const { responses } = await api.esQuery("structure", { size: ES_NO_LIMIT, query: { ids: { type: "_doc", values: structureIds } } });
-      if (responses?.length) {
-        const structures = responses[0]?.hits?.hits.map((e) => ({ _id: e._id, ...e._source }));
-        all = all.map((item) => ({ ...item, structure: structures?.find((e) => e._id === item.structureId) }));
-      }
-    }
     return all.map((data) => {
       if (!data.domains) data.domains = [];
       if (!data.structure) {
@@ -163,54 +145,6 @@ export default function List() {
 
   async function transformCandidature(data, selectedFields) {
     let all = data;
-
-    // Add tutor info
-    if (selectedFields.includes("missionTutor")) {
-      const tutorIds = [...new Set(data.map((item) => item.tutorId).filter((e) => e))];
-      if (tutorIds?.length) {
-        const queryTutor = {
-          query: { ids: { type: "_doc", values: tutorIds } },
-          track_total_hits: true,
-          size: ES_NO_LIMIT,
-        };
-        const resultTutor = await api.post(`/es/referent/export`, {
-          ...queryTutor,
-          fieldsToExport: missionCandidatureExportFields.find((f) => f.id === "missionTutor")?.fields,
-        });
-        if (resultTutor?.data?.length) {
-          all = data.map((item) => ({ ...item, tutor: resultTutor?.data?.find((e) => e._id === item.tutorId) }));
-        }
-      }
-    }
-
-    // Add structure info
-    let structureCategorie = ["structureInfo", "structureLocation"];
-    if (structureCategorie.some((e) => selectedFields.includes(e))) {
-      const structureIds = [...new Set(data.map((item) => item.structureId).filter((e) => e))];
-      const queryStructure = {
-        query: { ids: { type: "_doc", values: structureIds } },
-        track_total_hits: true,
-        size: ES_NO_LIMIT,
-      };
-
-      let fieldsToExportsStructure = [];
-
-      selectedFields.forEach((selected) => {
-        if (structureCategorie.includes(selected)) {
-          let fields = missionCandidatureExportFields.find((f) => f.id === selected)?.fields;
-          fieldsToExportsStructure = [...fieldsToExportsStructure, ...fields];
-        }
-      });
-
-      const resultStructure = await api.post(`/es/structure/export`, {
-        ...queryStructure,
-        fieldsToExport: fieldsToExportsStructure,
-      });
-
-      if (resultStructure?.data?.length) {
-        all = all.map((item) => ({ ...item, structure: resultStructure?.data?.find((e) => e._id === item.structureId) }));
-      }
-    }
 
     let youngCategorie = ["representative2", "representative1", "location", "address", "imageRight", "contact", "identity", "status"];
     let fieldsToExportsYoung = [];