+++ analytics/src/sentry.js
@@ -28,7 +28,7 @@ function initSentry(app) {
 function capture(err, contexte) {
   console.log("capture", err);
   if (!err) {
-    captureMessage("Error not defined");
+    sentryCaptureMessage("Error not defined");
     return;
   }
 
@@ -45,7 +45,7 @@ function capture(err, contexte) {
 function captureMessage(mess, contexte) {
   console.log("captureMessage", mess);
   if (!mess) {
-    captureMessage("Error not defined");
+    sentryCaptureMessage("Error not defined");
     return;
   }